refactored the rewardfragment viewmodel and viewmodel test
