+++ app/src/main/res/layout/fragment_rewards.xml
@@ -4,6 +4,7 @@
   xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:app="http://schemas.android.com/apk/res-auto"
   xmlns:tools="http://schemas.android.com/tools"
+  android:minHeight="120dp"
   android:layout_width="match_parent"
   android:layout_height="match_parent">
