+++ app/src/main/res/layout/project_layout.xml
@@ -58,12 +58,8 @@
     android:id="@+id/rewards_container"
     android:layout_width="match_parent"
     android:layout_height="0dp"
-    android:layout_alignParentBottom="true"
-    app:layout_constraintBottom_toBottomOf="parent"
-    android:elevation="@dimen/grid_2"
-    android:visibility="gone"
     app:cardBackgroundColor="@color/white"
-    app:cardCornerRadius="@dimen/grid_5"
+    app:cardCornerRadius="30dp"
     app:layout_constrainedHeight="false"
     app:layout_constraintTop_toTopOf="@+id/guideline">
 
@@ -90,13 +86,20 @@
 
       </com.google.android.material.appbar.AppBarLayout>
 
-      <fragment
-        android:name="com.kickstarter.ui.fragments.RewardsFragment"
+      <FrameLayout
         android:id="@+id/fragment_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-      </fragment>
+        <fragment
+          android:id="@+id/fragment_rewards"
+          android:name="com.kickstarter.ui.fragments.RewardsFragment"
+          tools:layout="@layout/fragment_rewards"
+          android:layout_width="match_parent"
+          android:layout_height="match_parent"/>
+
+      </FrameLayout>
+
     </LinearLayout>
 
     <com.google.android.material.button.MaterialButton
